# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cardre v2 API",
    "description": "Auditable open-source credit scorecard builder \u2014 v2.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "description": "Get a single project by ID.",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "description": "List all runs for a project.",
        "operationId": "list_runs_projects__project_id__runs_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "description": "Create and optionally execute a run for a plan version.",
        "operationId": "create_run_projects__project_id__runs_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "description": "Get a single run by ID with summary info.",
        "operationId": "get_run_projects__project_id__runs__run_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/runs/{run_id}/steps": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Run Steps",
        "description": "List all steps for a run.",
        "operationId": "list_run_steps_projects__project_id__runs__run_id__steps_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunStepResponse"
                  },
                  "title": "Response List Run Steps Projects  Project Id  Runs  Run Id  Steps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/runs/{run_id}/evidence": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Run Evidence",
        "description": "List all evidence edges for a run.",
        "operationId": "list_run_evidence_projects__project_id__runs__run_id__evidence_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunEvidenceEdgeResponse"
                  },
                  "title": "Response List Run Evidence Projects  Project Id  Runs  Run Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvidenceArtifactResponse": {
        "properties": {
          "evidence_artifact_id": {
            "type": "string",
            "title": "Evidence Artifact Id"
          },
          "evidence_edge_id": {
            "type": "string",
            "title": "Evidence Edge Id"
          },
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "evidence_artifact_id",
          "evidence_edge_id",
          "artifact_id",
          "role"
        ],
        "title": "EvidenceArtifactResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "cardre_version": {
            "type": "string",
            "title": "Cardre Version"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "name",
          "created_at",
          "cardre_version"
        ],
        "title": "ProjectResponse"
      },
...
      "RunCreateRequest": {
        "properties": {
          "plan_version_id": {
            "type": "string",
            "title": "Plan Version Id"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "sync": {
            "type": "boolean",
            "title": "Sync",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "plan_version_id"
        ],
        "title": "RunCreateRequest"
      },
      "RunEvidenceEdgeResponse": {
        "properties": {
          "evidence_edge_id": {
            "type": "string",
            "title": "Evidence Edge Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "run_step_id": {
            "type": "string",
            "title": "Run Step Id"
          },
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "parent_step_id": {
            "type": "string",
            "title": "Parent Step Id"
          },
          "policy": {
            "type": "string",
            "title": "Policy"
          },
          "source_label": {
            "type": "string",
            "title": "Source Label"
          },
          "is_reused": {
            "type": "boolean",
            "title": "Is Reused",
            "default": false
          },
          "is_stale": {
            "type": "boolean",
            "title": "Is Stale",
            "default": false
          },
          "stale_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale Reason"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/EvidenceArtifactResponse"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "evidence_edge_id",
          "run_id",
          "run_step_id",
          "step_id",
          "parent_step_id",
          "policy",
          "source_label"
        ],
        "title": "RunEvidenceEdgeResponse",
        "description": "Typed response for a single evidence edge in a run (#216).\n\nIncludes nested artifacts so the generated schema and frontend\ntypes match the actual payload shape."
      },
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunResponse"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "runs"
        ],
        "title": "RunListResponse"
      },
      "RunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "plan_version_id": {
            "type": "string",
            "title": "Plan Version Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "step_count": {
            "type": "integer",
            "title": "Step Count",
            "default": 0
          },
          "branch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "executed_step_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Executed Step Ids"
          },
          "diagnostics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "latest_error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Error"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "is_stale": {
            "type": "boolean",
            "title": "Is Stale",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "plan_version_id",
          "status",
          "started_at"
        ],
        "title": "RunResponse"
      },
      "RunStepResponse": {
        "properties": {
          "run_step_id": {
            "type": "string",
            "title": "Run Step Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "plan_version_id": {
            "type": "string",
            "title": "Plan Version Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "execution_fingerprint": {
            "additionalProperties": true,
            "type": "object",
            "title": "Execution Fingerprint"
          },
          "warnings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Warnings"
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "run_step_id",
          "run_id",
          "step_id",
          "plan_version_id",
          "status",
          "started_at"
        ],
        "title": "RunStepResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}